Integrate the bot into the main codebase
